feat: Wrapping scripts into generated HTML file
We now automatically generate a HTML wrapper for each script loaded.
Before, the scripts were loaded from a sandboxed iframe. The sandbox was hard to work with, especially forbidding CORS calls. The wrapper will be hosted along the map, usually in the map-storage where it has access to almost nothing from a security standpoint (the map-storage domain has no important cookies for instance). In the case of self-hosting in a single domain, we can assume the maps are uploaded by the administrator, so we don't need to protect ourselves agains rogue scripts from the maps.